clarify package/class name regex in ImportClass()
Consider the following line of Haxe code:

    setFacingFlip(FlxObject.LEFT, false, false);

Assuming the FlxObject class was not imported, a compilation error would be thrown. One might want to call the vaxe#ImportClass() function to remedy this issue. However, the previous implementation of the regex would match "FacingFlip", which is not the proper class name to search for in the ctags file.

In reality, package/class names must necessarily follow a non-word character in order to compile correctly. So, that constraint can be used to clarify the regex and more often guarantee a proper match.